refactor: Java Security Ultimate Scan 2023 (#54)
Disclaimer: Automated Commit Alert

Please be aware that this commit, generated through automated processes, may contain false alerts or not be precisely targeted. This automated commit is part of a large-scale effort to enhance software security over time. It is sent to various repositories to improve code quality and security. Exercise caution when reviewing the changes, and ensure that any necessary adjustments are made to maintain the integrity and functionality of the software.

Use this link to re-run the recipe: https://app.moderne.io/recipes/builder/MtR7dypUE?organizationId=SmVua2lucyBDSQ%3D%3D

Co-authored-by: Moderne <team@moderne.io>